Move 'Item list' component and styling to a separate folder.
