+++ client/my-sites/plans/jetpack-plans/product-store/item-list/index.tsx
@@ -1,6 +1,8 @@
-import { BundlesList } from './bundles-list';
-import { ProductsList } from './products-list';
-import type { ItemsListProps } from './types';
+import { BundlesList } from '../bundles-list';
+import { ProductsList } from '../products-list';
+import type { ItemsListProps } from '../types';
+
+import './style.scss';
 
 export const ItemsList: React.FC< ItemsListProps > = ( {
 	createCheckoutURL,
