+++ client/my-sites/plans/jetpack-plans/product-store/item-list/style.scss
@@ -1,0 +1,5 @@
+.jetpack-product-store__items-list .button.is-primary {
+	color: var( --color-text-inverted );
+	min-width: 5.7em;
+	border: none;
+}
